banner redirection + separating steps logic on welcome page
